backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
supabase>=2.3.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.24.0
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
numpy>=1.24.0
umap-learn>=0.5.5
PyPDF2>=3.0.1
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
groq>=0.4.0

